Fix issue where insertion could be aborted due to incorrect selection
If `this.selectedElementId` contains a value, but no element with that ID can be found in the editor, the insert will just break. Usually when this happens, it's the result of an element having been removed, without also properly clearing `this.selectedElementId`.

This PR ensures that if `this.selectedElementId` points to nothing, the insert will continue, and create a new embedded asset at the cursor position, as expected.