Fix redirect to image editor when no valid placement has been selected
In a previous "fix", I changed the behavior of the meta editor to send
the user back to the image editor after saving the metadata. This was
due to a complaint from several of the journalists that changing the
metadata along with adjustments to the image required too many clicks.

This approach does not work in the edge case where the user has not
selected a valid placement for the image. In that case, they should
only have access to altering the metadata, and be sent back to
DrPublish once it has been saved. I've added a conditional check to
either send them back to DrPublish or back to the image editor depending
on whether or not a valid selection has been made. I've also added
a warning to the user when the selection is invalid, so they know
what to do next.