Fix issue where images could be inserted without proper classes
There were two separate approached for inserting images, depending on whether or not you were replacing an image or inserting a new one. This is unnecessary, as DrPublish will handle replacement for you. One approach set the necessary classes, so that DrPublish can recognize it as an image, the other didn't. Simplified the logic to insert images in the same manner regardless of whether it's an update or not.